refactor: client 네트워크 요청 파일 제거
