+++ apis/news.ts
@@ -3,7 +3,7 @@
 import { News, NewsPreviewList } from '@/types/news';
 import { PostSearchQueryParams } from '@/types/post';
 
-import { deleteRequest, getRequest, patchRequest, postRequest } from './network/server';
+import { deleteRequest, getRequest, patchRequest, postRequest } from '.';
 
 const newsPath = '/news';
