+++ apis/seminar.ts
@@ -3,7 +3,7 @@
 import { PostSearchQueryParams } from '@/types/post';
 import { Seminar, SeminarList } from '@/types/seminar';
 
-import { postRequest, patchRequest, getRequest, deleteRequest } from './network/server';
+import { postRequest, patchRequest, getRequest, deleteRequest } from '.';
 
 const seminarPath = '/seminar';
