+++ constants/network.ts
@@ -1,4 +1,4 @@
-import { BASE_URL } from '@/apis/network/common';
+import { BASE_URL } from '@/apis';
 
 export const COOKIE_SESSION_ID = 'JSESSIONID';
 export const LOGIN_URL = BASE_URL + '/login';
